requirements.txt
babel==2.17.0
backrefs==5.9
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cssselect2==0.8.0
defusedxml==0.8.0rc2
ghp-import==2.1.0
idna==3.10
Jinja2==3.1.6
Markdown==3.8.2
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.18
mkdocs-material-extensions==1.3.1
packaging==25.0
paginate==0.5.7
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pycparser==2.22
Pygments==2.19.2
pymdown-extensions==10.16.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyyaml_env_tag==1.1
requests==2.32.5
six==1.17.0
tinycss2==1.4.0
urllib3==2.5.0
watchdog==6.0.0
webencodings==0.5.1
pillow==11.3.0
